+++ apps/backend/backend_plugin.c
@@ -114,6 +114,7 @@ clixon_plugin_statedata(clicon_handle    h,
     cxobj          *x = NULL;
     clixon_plugin  *cp = NULL;
     plgstatedata_t *fn;          /* Plugin statedata fn */
+    cxobj          *xerr = NULL;
     
     while ((cp = clixon_plugin_each(h, cp)) != NULL) {
 	if ((fn = cp->cp_api.ca_statedata) == NULL)
@@ -124,6 +125,22 @@ clixon_plugin_statedata(clicon_handle    h,
 	    goto fail;  /* Dont quit here on user callbacks */
 	if (xml_apply(x, CX_ELMNT, xml_spec_populate, yspec) < 0)
 	    goto done;
+	/* Check XML from state callback by validating it. return internal 
+	 * error with error cause 
+	 */
+	if ((ret = xml_yang_validate_all_top(h, x, &xerr)) < 0) 
+	    goto done;
+	if (ret > 0 && (ret = xml_yang_validate_add(h, x, &xerr)) < 0)
+	    goto done;
+	if (ret == 0){
+	    cbuf  *cberr = NULL;
+	    if (netconf_err2cb(xpath_first(xerr, "rpc-error"), &cberr) < 0)
+		goto done;
+	    clicon_log(LOG_WARNING, "%s: Internal error: state callback returned invalid XML: %s", __FUNCTION__, cbuf_get(cberr));
+	    if (cberr)
+		cbuf_free(cberr);
+	    /* Dont fail just log warning */
+	}
 	if ((ret = netconf_trymerge(x, yspec, xret)) < 0)
 	    goto done;
 	if (ret == 0)
@@ -135,6 +152,8 @@ clixon_plugin_statedata(clicon_handle    h,
     }
     retval = 1;
  done:
+   if (xerr)
+       xml_free(xerr);
     if (x)
 	xml_free(x);
     return retval;